PixelCPEGeneric modified to read params from cff

+++ RecoLocalTracker/SiPixelRecHits/interface/PixelCPEGeneric.h
@@ -121,8 +121,8 @@ class PixelCPEGeneric : public PixelCPEBase
   double EdgeClusterErrorX_;
   double EdgeClusterErrorY_;
 
-  std::vector<float> xerr_barrel_l1_,yerr_barrel_l1_,xerr_barrel_ln_;
-  std::vector<float> yerr_barrel_ln_,xerr_endcap_,yerr_endcap_;
+  std::vector<double> xerr_barrel_l1_,yerr_barrel_l1_,xerr_barrel_ln_;
+  std::vector<double> yerr_barrel_ln_,xerr_endcap_,yerr_endcap_;
   float xerr_barrel_l1_def_, yerr_barrel_l1_def_,xerr_barrel_ln_def_;
   float yerr_barrel_ln_def_, xerr_endcap_def_, yerr_endcap_def_;
